Phase 4.4: Fix sign-compare warning in spew.c - size_t type modernization
Problem: 1 sign-compare warning comparing int with size_t
- spew.c:1020: spec_len < sizeof(temp_spec)

Root Cause: Variable 'spec_len' declared as int but compared with size_t from sizeof()

Solution Applied:
1. Changed spec_len declaration from 'int spec_len' to 'size_t spec_len'
2. Maintains consistent type usage for memory/size operations
3. No functional changes - same values, proper types

Technical Analysis:
- size_t is the proper type for memory size operations and sizeof() results
- Eliminates signed/unsigned comparison warnings
- Follows established Phase 4.4 pattern: use correct types for system interfaces
- Similar fix pattern to admin.c uid_t modernization

Validation:
- Compilation test: gcc -Wall -Wextra spew.c → zero warnings
- Sign-compare warning eliminated
- No regressions introduced

Progress: Phase 4.4 spew.c complete (5/47 warnings fixed)

+++ spew.c
@@ -1016,7 +1016,7 @@ static void generate_text(const char *class_spec, char default_variant, FILE *ou
                     if (*p == DELIMITER_CHAR) {
                         p += 2; /* Skip delimiter and variant tag */
                         char temp_spec[64];
-                        int spec_len = (p - start);
+                        size_t spec_len = (p - start);
                         if (spec_len < sizeof(temp_spec)) {
                             memcpy(temp_spec, start, spec_len);
                             temp_spec[spec_len] = '\0';
